requirements.txt
Flask==3.1.2
pytest==8.2.0
requests==2.32.5
allure-pytest==2.15.0
PyYAML==6.0.3